requirements.txt
flask==3.0.3
flask-cors==4.0.1
groq==0.9.0
chromadb==0.5.3
sentence-transformers==3.0.1
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
werkzeug==3.0.3
tiktoken==0.7.0
numpy==1.26.4
httpx==0.27.2